Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/url/004-expected.txt


FAIL WebSockets: 'URL' assert_equals: expected (undefined) undefined but got (string) "ws://localhost:49001/"
